auth-service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
sqlalchemy
aiosqlite
